Renombrar la clase de RutinaActivity.kt como Nutricion.kt y conectarla con la NavegacionApp.kt
Correcion de la navegacion entre pantallas, Ahora isguen una ruta mas logica

p.e Rutinas lleva a Rutina Semanal
Perfil lleva a lo que se puso en el test Inicial

Y dentro de Ahi se puede acceder a la rutina semanal y al Plan Alimenticio con Nutricion
