+++ app/src/main/java/com/example/ryuu_fit/pantallas/detalles_training.kt
@@ -20,11 +20,14 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.example.ryuu_fit.R
 
+
 // Modelo de datos para cada ejercicio
 data class Ejercicio(
     val nombre: String,
     val repeticiones: String,
-    val imagen: Int
+    val imagen: Int,
+    //val navController: NavController,
+    //val onClick: () -> Unit
 )
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -38,7 +41,7 @@ fun detallesTraining(
     val ejercicios = listOf(
         Ejercicio("Flexiones en Pica", "4×12", R.drawable.flexion_pica), // flexion_pica
         Ejercicio("Plancha Frontal", "3×45", R.drawable.plancha_frontal), // plancha_frontal
-        Ejercicio("Abdominales en Bicicleta", "3×20", R.drawable.abd_bici), // abd_bici
+        Ejercicio("Abdominales en Bicicleta", "3×20", R.drawable.abd_bici), // abd_bici  onClick = { navController.navigate("detallesEx")}
         Ejercicio("Elevaciones Laterales", "3×12", R.drawable.elevaciones_laterales) // elevaciones_laterales
     )
